ZEB-573 finish slider/refactoring file structure in tutorials module
