+++ client/app/modules/tutorials/data.js
@@ -1,0 +1,37 @@
+export default {
+    tutorials: [
+        {
+            route: 'dashboard',
+            steps: [
+                {
+                    order: 0,
+                    title: 'T1',
+                    description: 'D1',
+                    id: 'gocwRvLhDf8',
+                },
+                {
+                    order: 1,
+                    title: 'T2',
+                    description: 'D2',
+                    id: 'TcHIZuSn-9I',
+                },
+                {
+                    order: 1,
+                    title: 'T2',
+                    description: 'D2',
+                    id: 'TcHIZuSn-9I',
+                },
+                {
+                    order: 1,
+                    title: 'T2',
+                    description: 'D2',
+                    id: 'TcHIZuSn-9I',
+                },
+            ],
+        },
+        {
+            route: 'users',
+            steps: [],
+        },
+    ],
+};
