+++ client/app/modules/tutorials/tutorials-modal/index.js
@@ -1,0 +1,6 @@
+import TutorialsModalTemplate from './tutorials-modal.component.html';
+
+export { TutorialsModalController } from './tutorials-modal.controller';
+export {
+    TutorialsModalTemplate,
+};
