+++ client/app/modules/tutorials/tutorials-slider/tutorials-slider.component.html
@@ -1,26 +1,29 @@
-<div class="tutorials-slider">
-    <div class="tutorials-slider__slides">
-        <div class="tutorials-slider__item-wrapper" ng-style="{transform: $ctrl.transform}">
+<div class="zeb-tutorials-slider">
+    <div class="zeb-tutorials-slider__slides">
+        <div class="zeb-tutorials-slider__item-wrapper" ng-style="{transform: $ctrl.transform}">
             <youtube ng-repeat="slide in $ctrl.slides track by $index"
-                     class="tutorials-slider__item"
+                     class="zeb-tutorials-slider__item"
                      video-id="slide.id"
                      width="100%"
                      height="auto"></youtube>
         </div>
 
-        <button class="tutorials-slider__arrow _left" ng-click="$ctrl.prev()">
+        <button class="zeb-tutorials-slider__arrow _left" ng-click="$ctrl.prev()">
             <md-icon class="material-icons">arrow_back</md-icon>
         </button>
-        <button class="tutorials-slider__arrow _right" ng-click="$ctrl.next()">
+        <button class="zeb-tutorials-slider__arrow _right" ng-click="$ctrl.next()">
             <md-icon class="material-icons">arrow_forward</md-icon>
         </button>
     </div>
 
-    <div class="tutorials-slider__dots">
-        <div class="tutorials-slider__dots-item"
+    <div class="zeb-tutorials-slider__dots">
+        <div class="zeb-tutorials-slider__dots-item"
              ng-repeat="slide in $ctrl.slides track by $index"
              ng-class="{_active: $index === $ctrl.slideIndex}"
              ng-click="$ctrl.changeSlide($index)">
         </div>
     </div>
+
+    <h2 class="zeb-tutorials-slider__title">{{ $ctrl.activeSlide.title }}</h2>
+    <p class="zeb-tutorials-slider__description">{{ $ctrl.activeSlide.description }}</p>
 </div>
