+++ client/app/modules/tutorials/tutorials-slider/tutorials-slider.component.scss
@@ -1,4 +1,4 @@
-.tutorials-slider {
+.zeb-tutorials-slider {
     &__dots {
         display: flex;
         justify-content: center;
@@ -13,15 +13,15 @@
             margin: 0 8px;
             cursor: pointer;
             &._active {
+                cursor: default;
                 &:after {
                     content: '';
                     position: absolute;
-                    width: 100%;
-                    height: 100%;
-                    left: 0;
-                    top: 0;
+                    width: 14px;
+                    height: 14px;
+                    left: -3px;
+                    top: -3px;
                     border: 1px solid #011627;
-                    transform: scale(1.75);
                     border-radius: 50%;
                 }
             }
@@ -82,5 +82,19 @@
         .material-icons {
             color: #011627;
         }
+    }
+
+    &__title {
+        font-weight: 700;
+        font-size: 24px;
+        line-height: 28px;
+        color: #011627;
+        margin-bottom: 8px;
+    }
+    &__description {
+        font-size: 14px;
+        line-height: 24px;
+        color: #011627;
+        margin: 0;
     }
 }
