+++ client/app/modules/tutorials/tutorials-slider/tutorials-slider.controller.js
@@ -4,19 +4,28 @@ export default function TutorialsSliderController() {
     return {
         slides: null,
         slideIndex: 0,
-        changeSlide,
+
+        get activeSlide() {
+            return this.slides && this.slides[this.slideIndex];
+        },
+
         get transform() {
             return `translateX(-${this.slideIndex * 100}%)`;
         },
 
-        prev() {
-            const newIndex = this.slideIndex - 1;
-            this.changeSlide(newIndex < 0 ? this.slides.length - 1 : newIndex);
-        },
-        next() {
-            this.changeSlide((this.slideIndex + 1) % this.slides.length);
-        },
+        changeSlide,
+        prev,
+        next,
     };
+
+    function prev() {
+        const newIndex = this.slideIndex - 1;
+        this.changeSlide(newIndex < 0 ? this.slides.length - 1 : newIndex);
+    }
+
+    function next() {
+        this.changeSlide((this.slideIndex + 1) % this.slides.length);
+    }
 
     function changeSlide(index) {
         this.slideIndex = index;
