+++ client/app/modules/tutorials/tutorials.styles.scss
@@ -1,0 +1,40 @@
+.zeb-tutorials-container {
+    position: fixed;
+    bottom: -8px;
+    z-index: 1000;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    left: 50%;
+    transform: translateX(-50%);
+}
+
+.zeb-tutorial {
+    transition: transform .2s ease-out;
+    background: #011627;
+    padding: 8px 12px 16px;
+    border-radius: 8px;
+    color: #fff;
+    font-size: 12px;
+    font-weight: 700;
+    transform: translateY(100%);
+    md-icon {
+        cursor: pointer;
+        color: #F2F2F2;
+        margin-left: 12px;
+        margin-right: -8px;
+    }
+    &.entered {
+        transform: translateY(0);
+    }
+    .zeb-tutorials-container._opened &,
+    .zeb-tutorials-container._hide:not(:hover) &{
+        transform: translateY(60%);
+    }
+}
+
+.zeb-tutorials-modal {
+    &__content {
+        padding: 24px 24px 16px;
+    }
+}
