+++ client/styles/app.scss
@@ -44,59 +44,3 @@
 ,   "components/modals/ci-helper-modal/ci-helper-modal"
 ,   "components/modals/review-modal/review-modal"
 ;
-
-
-.zeb-tutorials-container {
-    position: fixed;
-    bottom: -8px;
-    z-index: 1000;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    left: 50%;
-    transform: translateX(-50%);
-    // transform: translateY(0);
-}
-.zeb-tutorial {
-    transition: transform .2s ease-out;
-    background: #011627;
-    padding: 8px 12px 16px;
-    border-radius: 8px;
-    color: #fff;
-    font-size: 12px;
-    font-weight: 700;
-    transform: translateY(100%);
-    md-icon {
-        cursor: pointer;
-        color: #F2F2F2;
-        margin-left: 12px;
-        margin-right: -8px;
-    }
-    &.entered {
-        transform: translateY(0);
-    }
-}
-.zeb-tutorials-container._hide:not(:hover) {
-    .zeb-tutorial {
-        transform: translateY(60%);
-    }
-}
-
-.tutorials-modal {
-    &__content {
-        padding: 24px 24px 16px;
-    }
-    &__title {
-        font-weight: 700;
-        font-size: 24px;
-        line-height: 28px;
-        color: #011627;
-        margin-bottom: 8px;
-    }
-    &__description {
-        font-size: 14px;
-        line-height: 24px;
-        color: #011627;
-        margin: 0;
-    }
-}
